实现用户进程hello, exit

+++ kernel/driver/ide.h
@@ -8,7 +8,7 @@
 
 #define IDE_SIZE        (SECTION_SIZE * IDE_SECTIONS)
 
-void read_ide_sects(void* dst, uint32_t src_sect, size_t num_sects);
-void write_ide_sects(const void* src, uint32_t dst_sect, size_t num_sects);
+void read_ide_sects(void* dst, u32 src_sect, u64 num_sects);
+void write_ide_sects(const void* src, u32 dst_sect, u64 num_sects);
 
 #endif // __KERNEL_DRIVER_IDE_H__
