+++ lib/string.c
@@ -2,17 +2,17 @@
 
 #include "defs.h"
 
-size_t strlen(const char* s) {
-    size_t cnt = 0;
+u64 strlen(const char* s) {
+    u64 cnt = 0;
     while (*s++ != '\0')
         cnt++;
     return cnt;
 }
 
-int32_t strcmp(const char* s1, const char* s2) {
+i32 strcmp(const char* s1, const char* s2) {
     while (*s1 && *s1 == *s2)
         s1++, s2++;
-    return (int32_t) (*s1 - *s2);
+    return (i32) (*s1 - *s2);
 }
 
 char* strcpy(char* dst, const char* src) {
@@ -23,8 +23,8 @@ char* strcpy(char* dst, const char* src) {
 }
 
 char* strrev(char* s) {
-    size_t len = strlen(s);
-    for (size_t i = 0; i < len / 2; i++) {
+    u64 len = strlen(s);
+    for (u64 i = 0; i < len / 2; i++) {
         char temp = s[i];
         s[i] = s[len - 1 - i];
         s[len - 1 - i] = temp;
@@ -32,16 +32,16 @@ char* strrev(char* s) {
     return s;
 }
 
-void* memset(void* s, char c, size_t n) {
+void* memset(void* s, char c, u64 n) {
     char* p = s;
     while (n-- > 0) 
         *p++ = c;
     return s;
 }
 
-void* memcpy(void* dst, const void* src, size_t n) {
-    const byte* s = src;
-    byte* p = dst;
+void* memcpy(void* dst, const void* src, u64 n) {
+    const u8* s = src;
+    u8* p = dst;
     while (n--) 
         *p++ = *s++;
     return dst;
